Move callbacks into com.google.audio
